fix chenges requested
