+++ index.html
@@ -22,7 +22,7 @@
       <div class="map__canvas" id="map-canvas"></div>
 
       <!-- Фильтрация объявлений -->
-      <div class="map__filters-container ad-form--disabled">
+      <div class="map__filters-container">
         <form action="#" class="map__filters" autocomplete="off">
           <select name="housing-type" id="housing-type" class="map__filter">
             <option value="any" selected>Любой тип жилья</option>
