+++ js/main.js
@@ -1,2 +1,3 @@
-import './modules/offers-generator.js';
-import './modules/fill-cards-template.js';
+import {generateOffers} from './modules/offers-generator.js';
+import {generateCards} from './modules/cards-generator.js';
+generateCards(generateOffers(2));
